pyproject.toml
[tool.poetry]
name = "pytest_sample"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
requests = "^2.32.3"
pytest = "^8.3.2"
faker = "^37.12.0"
allure-pytest = "^2.13.5"
pytest-dotenv = "^0.5.2"
confluent-kafka = "~2.5.0"
jsonschema = "^4.23.0"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
playwright = "^1.52.0"
pytest-rerunfailures = "^16.1"
boto3 = "^1.40.6"
pytest-xdist = "^3.8.0"
pydantic = "^2.12.5"
pydantic-settings = "^2.13.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
poethepoet = "^0.26.1"
isort = "^5.13.2"
pre-commit = "^3.7.1"
flask = "^3.1.0"

# Run tests
[tool.poe.tasks]
test_task_api = "pytest -m \"task_api\" -vv"

# Linting and formating
[tool.poe.tasks.format]
cmd = "black ."

[tool.poe.tasks.sort]
cmd = "isort ."

[tool.poe.tasks.lint]
cmd = "flake8 ."

# Reports
[tool.poe.tasks.report]
cmd = "allure serve allure-results"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '\.pyi?$'
line-length = 120

[tool.isort]
skip = [".gitignore", ".venv", "allure-results"]
profile = "black"
line_length = 120
